fix: infer local issue links on workspace start

+++ crates/db/src/models/scratch.rs
@@ -143,12 +143,14 @@ pub struct UiPreferencesData {
     pub selected_project_id: Option<String>,
     /// Preferred order for local-only project icons in the app bar
     #[serde(default)]
+    #[ts(skip)]
     pub local_project_order: Vec<String>,
     /// Default setting for creating a draft workspace from new issues
     #[serde(default)]
     pub create_draft_workspace_by_default: Option<bool>,
     /// Whether to show optional links and profile controls in the left app bar
     #[serde(default)]
+    #[ts(skip)]
     pub show_left_column_links: Option<bool>,
     /// Kanban project view selections (active view per project)
     #[serde(default)]
@@ -206,7 +208,7 @@ pub struct ProjectStatusConfigData {
     pub color: String,
     #[serde(default)]
     pub hidden: bool,
-    pub sort_order: i64,
+    pub sort_order: i32,
 }
 
 /// Data for project repo defaults scratch (default repos/branches per project)
@@ -250,6 +252,7 @@ pub struct DraftIssueData {
 #[strum_discriminants(ts(use_ts_enum))]
 #[strum_discriminants(serde(rename_all = "SCREAMING_SNAKE_CASE"))]
 #[strum_discriminants(strum(serialize_all = "SCREAMING_SNAKE_CASE"))]
+#[allow(clippy::large_enum_variant)]
 pub enum ScratchPayload {
     DraftTask(String),
     DraftFollowUp(DraftFollowUpData),
